Выполнить финальный проект спринта 11 исправить замечания ревьюера
